+++ Mitosis/ICloneEngine.cs
@@ -1,3 +1,5 @@
+using System.Diagnostics.CodeAnalysis;
+
 namespace Nanoray.Mitosis;
 
 /// <summary>
@@ -6,10 +8,31 @@ namespace Nanoray.Mitosis;
 public interface ICloneEngine
 {
 	/// <summary>
-	/// Clones the given value recursively.
+	/// Attempts to clone the given value.
 	/// </summary>
-	/// <param name="value">The value to clone.</param>
+	/// <param name="original">The value to clone.</param>
+	/// <param name="clone">The cloned value, if succeeded.</param>
+	/// <typeparam name="T">The type of value to clone.</typeparam>
+	/// <returns>Whether cloning was successful.</returns>
+	bool TryClone<T>(T original, [MaybeNullWhen(false)] out T clone)
+	{
+		try
+		{
+			clone = this.Clone(original);
+			return true;
+		}
+		catch
+		{
+			clone = default;
+			return false;
+		}
+	}
+	
+	/// <summary>
+	/// Clones the given value.
+	/// </summary>
+	/// <param name="original">The value to clone.</param>
 	/// <typeparam name="T">The type of value to clone.</typeparam>
 	/// <returns>The cloned value.</returns>
-	T Clone<T>(T value);
+	T Clone<T>(T original);
 }